CRS-2493: Refactor calc summary
Move to a controller, this enables use of Zod to validate the "Do you agree with these dates?" question which is to be added for genuine overrides.
It is currently hardcoded to YES within the form.
