+++ server/routes/calculationRoutes.test.ts
@@ -1,7 +1,5 @@
 import request from 'supertest'
 import type { Express } from 'express'
-import { HttpError } from 'http-errors'
-import MockDate from 'mockdate'
 import * as cheerio from 'cheerio'
 import { appWithAllRoutes } from './testutils/appSetup'
 import PrisonerService from '../services/prisonerService'
@@ -18,7 +16,6 @@ import {
 } from '../@types/calculateReleaseDates/calculateReleaseDatesClientTypes'
 import ReleaseDateWithAdjustments from '../@types/calculateReleaseDates/releaseDateWithAdjustments'
 import UserInputService from '../services/userInputService'
-import { pedAdjustedByCrdAndBeforePrrdBreakdown } from '../services/breakdownExamplesTestData'
 import ViewReleaseDatesService from '../services/viewReleaseDatesService'
 import { expectMiniProfile } from './testutils/layoutExpectations'
 import { ResultsWithBreakdownAndAdjustments } from '../@types/calculateReleaseDates/rulesWithExtraAdjustments'
@@ -470,320 +467,6 @@ describe('Calculation routes tests', () => {
         expect(res.text).toContain("Please select either 'Yes' or 'No'")
       })
   })
-  it('GET /calculation/A1234AB/123456/confirmation should return details about the calculation requested', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue(
-      stubbedResultsWithBreakdownAndAdjustments,
-    )
-    return request(app)
-      .get('/calculation/A1234AB/123456/confirmation')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        const $ = cheerio.load(res.text)
-        expect($('[data-qa=cancel-link]').first().attr('href')).toStrictEqual(
-          '/calculation/A1234AA/cancelCalculation?redirectUrl=/calculation/A1234AB/123456/confirmation',
-        )
-        const submitToNomis = $('[data-qa=submit-to-nomis]').first()
-        expect(submitToNomis.length).toStrictEqual(1)
-      })
-  })
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should return details about the calculation requested', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue(
-      stubbedResultsWithBreakdownAndAdjustments,
-    )
-    return request(app)
-      .get('/calculation/A1234AB/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        expect(res.text).toContain('CRD')
-        expect(res.text).toContain('Conditional release date')
-        expect(res.text).toContain('Wednesday, 03 February 2021')
-        expect(res.text).toContain('Tuesday, 02 February 2021 when adjusted to a working day')
-        expect(res.text).toContain('HDCED')
-        expect(res.text).toContain('Home detention curfew eligibility date')
-        expect(res.text).toContain('Sunday, 03 October 2021')
-        expect(res.text).toContain('Tuesday, 05 October 2021 when adjusted to a working day')
-        // expect(res.text).not.toContain('SLED')
-        // This is now displayed as part of breakdown even IF the dates don't contain a SLED.
-        // The design without SLED will come in time
-        expect(res.text).toContain('Sentence')
-        expect(res.text).not.toContain('Consecutive sentence')
-        expect(res.text).toContain('Release dates with adjustments')
-        expect(res.text).toContain('03 February 2021')
-        expect(res.text).toContain('15 January 2021 minus 18 days')
-        expect(res.text).toContain('HDCED with adjustments')
-        expect(res.text).toContain('13 May 2029')
-        expect(res.text).toContain('14 May 2029 minus 1 day')
-        expect(res.text).toContain(
-          `Some release dates and details are not included because they are not relevant to this person's sentences`,
-        )
-        expect(res.text).toContain(`Monday, 03 February 2020`)
-        expect(res.text).toContain(`ERSED`)
-        expect(res.text).toContain('Early removal scheme eligibility date')
-        expect(res.text).not.toContain('From 16 January, the policy for calculating ERSED has changed')
-        expectMiniProfile(res.text, expectedMiniProfile)
-        expect(res.text).not.toContain(
-          'Early removal cannot happen as release from the Detention Training Order (DTO) is later than the Conditional Release Date (CRD).',
-        )
-        expect(res.text).toContain('Calculation breakdown')
-        const $ = cheerio.load(res.text)
-        expect($('[data-qa=cancel-link]').first().attr('href')).toStrictEqual(
-          '/calculation/A1234AA/cancelCalculation?redirectUrl=/calculation/A1234AB/summary/123456',
-        )
-        const submitToNomis = $('[data-qa=submit-to-nomis]').first()
-        expect(submitToNomis.attr('href')).toStrictEqual('/calculation/A1234AA/123456/approved-dates-question')
-      })
-  })
-
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should hide breakdown if feature toggle is off', () => {
-    config.featureToggles.showBreakdown = false
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue(
-      stubbedResultsWithBreakdownAndAdjustments,
-    )
-    return request(app)
-      .get('/calculation/A1234AB/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        expect(res.text).not.toContain('Calculation breakdown')
-      })
-  })
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should show ERSED recall notification banner if recall only', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue({
-      ...stubbedResultsWithBreakdownAndAdjustments,
-      calculationOriginalData: {
-        ...stubbedResultsWithBreakdownAndAdjustments,
-        sentencesAndOffences: [
-          {
-            bookingId: 1,
-            sentenceSequence: 3,
-            lineSequence: 3,
-            caseSequence: 3,
-            courtDescription: 'Preston Crown Court',
-            sentenceStatus: 'A',
-            sentenceCategory: '2020',
-            sentenceCalculationType: 'LR_EDS18',
-            sentenceTypeDescription: 'LR_EDS18',
-            sentenceDate: '2021-09-03',
-            terms: [
-              {
-                years: 0,
-                months: 2,
-                weeks: 0,
-                days: 0,
-                code: 'IMP',
-              },
-            ],
-            offence: {
-              offenderChargeId: 1,
-              offenceStartDate: '2020-01-01',
-              offenceCode: 'RL05016',
-              offenceDescription: 'Access / exit by unofficial route - railway bye-law',
-              indicators: [],
-            },
-            isSDSPlus: false,
-            hasAnSDSEarlyReleaseExclusion: 'NO',
-            isSDSPlusEligibleSentenceTypeLengthAndOffence: false,
-            isSDSPlusOffenceInPeriod: false,
-          },
-        ],
-      },
-    })
-    return request(app)
-      .get('/calculation/A1234AB/summary/123456')
-      .expect(200)
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        expect(res.text).not.toContain('Include an Early removal scheme eligibility date (ERSED) in this calculation')
-        expect(res.text).toContain('Important')
-        expect(res.text).toContain(
-          'This service cannot calculate the ERSED if the person is serving a recall. If they are eligible for early removal, enter the ERSED in NOMIS.',
-        )
-      })
-  })
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should not blow up if breakdown is missing', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue({
-      ...stubbedResultsWithBreakdownAndAdjustments,
-      calculationBreakdown: undefined,
-      releaseDatesWithAdjustments: undefined,
-      breakdownMissingReason: 'UNSUPPORTED_CALCULATION_BREAKDOWN',
-    })
-
-    return request(app)
-      .get('/calculation/A1234AB/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        expect(res.text).toContain('CRD')
-        expect(res.text).toContain('Conditional release date')
-        expect(res.text).toContain('Wednesday, 03 February 2021')
-        expect(res.text).toContain('Tuesday, 02 February 2021 when adjusted to a working day')
-        expectMiniProfile(res.text, expectedMiniProfile)
-      })
-  })
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should show hints generated by the API', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue({
-      ...stubbedResultsWithBreakdownAndAdjustments,
-      context: { ...stubbedResultsWithBreakdownAndAdjustments.context, prisonerId: 'A1234AA' },
-      calculationBreakdown: pedAdjustedByCrdAndBeforePrrdBreakdown(),
-      dates: {
-        SLED: { date: '2029-09-14', type: 'SLED', description: 'Sentence and licence expiry date', hints: [] },
-        CRD: { date: '2026-09-14', type: 'CRD', description: 'Conditional release date', hints: [] },
-        PED: {
-          date: '2024-10-12',
-          type: 'PED',
-          description: 'Parole eligibility date',
-          hints: [
-            { text: 'PED adjusted for the CRD of a concurrent sentence or default term' },
-            { text: 'The post recall release date (PRRD) of Tuesday, 18 March 2025 is later than the PED' },
-          ],
-        },
-        ESED: { date: '2029-09-14', type: 'ESED', description: 'Effective sentence end date', hints: [] },
-      },
-    })
-    return request(app)
-      .get('/calculation/A1234AA/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        expect(res.text).toContain('PED adjusted for the CRD of a concurrent sentence or default term')
-        expect(res.text).toContain(
-          'The post recall release date (PRRD) of Tuesday, 18 March 2025 is later than the PED',
-        )
-        expect(res.text).not.toContain('Important')
-        expect(res.text).not.toContain(
-          'This service cannot calculate the ERSED if the person is serving a recall. If they are eligible for early removal, enter the ERSED in NOMIS.',
-        )
-      })
-  })
-
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should display notification when ERSED cannot happen because of DTO', () => {
-    const stubbedCalculationBreakdownWithErsedBanner: CalculationBreakdown = {
-      showSds40Hints: false,
-      concurrentSentences: [
-        {
-          dates: {
-            CRD: {
-              adjusted: '2021-02-03',
-              unadjusted: '2021-01-15',
-              adjustedByDays: 18,
-              daysFromSentenceStart: 100,
-            },
-            SED: {
-              adjusted: '2021-02-03',
-              unadjusted: '2021-01-15',
-              adjustedByDays: 18,
-              daysFromSentenceStart: 100,
-            },
-          },
-          sentenceLength: '2 years',
-          sentenceLengthDays: 785,
-          sentencedAt: '2020-01-01',
-          lineSequence: 2,
-          caseSequence: 1,
-          externalSentenceId: {
-            sentenceSequence: 0,
-            bookingId: 0,
-          },
-        },
-      ],
-      breakdownByReleaseDateType: {},
-      otherDates: {},
-      ersedNotApplicableDueToDtoLaterThanCrd: true,
-    }
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue({
-      ...stubbedResultsWithBreakdownAndAdjustments,
-      context: { ...stubbedResultsWithBreakdownAndAdjustments.context, prisonerId: 'A1234AA' },
-      dates: {
-        SLED: { date: '2023-09-20', type: 'SLED', description: 'Sentence and licence expiry date', hints: [] },
-        MTD: { date: '2024-12-20', type: 'MTD', description: 'Mid transfer date', hints: [] },
-        ERSED: { date: '2021-12-20', type: 'ERSED', description: 'Early removal scheme eligibility date', hints: [] },
-        CRD: { date: '2022-08-14', type: 'CRD', description: 'Conditional release date', hints: [] },
-        ESED: { date: '2023-09-20', type: 'ESED', description: 'Effective sentence end date', hints: [] },
-      },
-      calculationBreakdown: stubbedCalculationBreakdownWithErsedBanner,
-    })
-
-    return request(app)
-      .get('/calculation/A1234AA/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        const $ = cheerio.load(res.text)
-        const ersedNABanner = $('[data-qa=ersed-na-banner]').first()
-        const impTitle = $('[data-qa=important-title]').first()
-        expect(impTitle.text()).toStrictEqual('Important')
-        expect(ersedNABanner.text()).toStrictEqual(
-          'Early removal cannot happen as release from the Detention Training Order (DTO) is later than the Conditional Release Date (CRD).',
-        )
-      })
-  })
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should not error and also not display ERSED Not applicable banner when calculationBreakdown is null', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue({
-      ...stubbedResultsWithBreakdownAndAdjustments,
-      context: { ...stubbedResultsWithBreakdownAndAdjustments.context, prisonerId: 'A1234AA' },
-      dates: {
-        SLED: { date: '2023-09-20', type: 'SLED', description: 'Sentence and licence expiry date', hints: [] },
-        MTD: { date: '2024-12-20', type: 'MTD', description: 'Mid transfer date', hints: [] },
-        ERSED: { date: '2021-12-20', type: 'ERSED', description: 'Early removal scheme eligibility date', hints: [] },
-        CRD: { date: '2022-08-14', type: 'CRD', description: 'Conditional release date', hints: [] },
-        ESED: { date: '2023-09-20', type: 'ESED', description: 'Effective sentence end date', hints: [] },
-      },
-      calculationBreakdown: null,
-    })
-
-    return request(app)
-      .get('/calculation/A1234AA/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        const $ = cheerio.load(res.text)
-        const ersedNABanner = $('[data-qa=ersed-na-banner]').first()
-        const impTitle = $('[data-qa=important-title]').first()
-        expect(impTitle.length).toBe(0)
-        expect(ersedNABanner.length).toBe(0)
-      })
-  })
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should display upcoming HDCED changes notification', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue(
-      stubbedResultsWithBreakdownAndAdjustments,
-    )
-    return request(app)
-      .get('/calculation/A1234AB/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        expect(res.text).not.toContain('From 6 June, the policy for calculating HDCED has changed')
-        expect(res.text).not.toContain('This service has calculated the HDCED using the new policy rules')
-      })
-  })
-
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should display HDCED changes notification', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue(
-      stubbedResultsWithBreakdownAndAdjustments,
-    )
-    MockDate.set('2023-06-06')
-    return request(app)
-      .get('/calculation/A1234AB/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        expect(res.text).not.toContain('This service has calculated the HDCED using the new policy rules.')
-        expect(res.text).not.toContain('From 6 June, the policy for calculating HDCED has changed')
-      })
-  })
 
   it('GET /calculation/:nomsId/complete should return details about the calculation requested', () => {
     prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
@@ -895,20 +578,6 @@ describe('Calculation routes tests', () => {
       })
   })
 
-  it('GET /calculation/:nomsId/summary should return confirm and continue button if approved dates on', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue({
-      ...stubbedResultsWithBreakdownAndAdjustments,
-      context: { ...stubbedResultsWithBreakdownAndAdjustments.context, prisonerId: 'A1234AA' },
-    })
-    return request(app)
-      .get('/calculation/A1234AA/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        expect(res.text).toContain('Confirm and continue')
-      })
-  })
   it('GET /calculation/:nomsId/complete should error if calculation is not confirmed', () => {
     prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
     calculateReleaseDatesService.getCalculationResults.mockResolvedValue({
@@ -918,41 +587,6 @@ describe('Calculation routes tests', () => {
     return request(app).get('/calculation/A1234AB/complete/123456').expect(404).expect('Content-Type', /html/)
   })
 
-  it('POST /calculation/:nomsId/summary/:calculationRequestId should redirect if an error is thrown', () => {
-    const error = {
-      status: 412,
-      message: 'An error has occurred',
-    } as HttpError
-
-    calculateReleaseDatesService.confirmCalculation.mockImplementation(() => {
-      throw error
-    })
-    return request(app)
-      .post('/calculation/A1234AB/summary/123456')
-      .expect(302)
-      .expect(res => {
-        expect(res.redirect).toBeTruthy()
-      })
-  })
-  it('POST /calculation/:nomsId/summary/:calculationRequestId should submit release dates', () => {
-    calculateReleaseDatesService.confirmCalculation.mockResolvedValue({
-      dates: {},
-      calculationRequestId: 654321,
-      effectiveSentenceLength: null,
-      prisonerId: 'A1234AA',
-      calculationReference: 'ABC123',
-      bookingId: 123,
-      calculationStatus: 'PRELIMINARY',
-      calculationType: 'CALCULATED',
-    })
-    return request(app)
-      .post('/calculation/A1234AB/summary/123456')
-      .expect(302)
-      .expect('Location', '/calculation/A1234AB/complete/654321')
-      .expect(res => {
-        expect(res.redirect).toBeTruthy()
-      })
-  })
   it('GET /calculation/:nomsId/summary/:calculationRequestId/print should return a printable page about the calculation requested', () => {
     prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
     calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue(
@@ -967,22 +601,6 @@ describe('Calculation routes tests', () => {
         expect(res.text).toMatch(/<script src="\/assets\/print.js"><\/script>/)
         expect(res.text).toMatch(/Dates for/)
         expectMiniProfile(res.text, expectedMiniProfile)
-      })
-  })
-
-  it('GET /calculation/:nomsId/summary/:calculationRequestId should not display tranche label', () => {
-    prisonerService.getPrisonerDetail.mockResolvedValue(stubbedPrisonerData)
-    const dataWithTranche1 = { ...stubbedResultsWithBreakdownAndAdjustments }
-    dataWithTranche1.tranche = 'TRANCHE_1'
-    calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments.mockResolvedValue(dataWithTranche1)
-    return request(app)
-      .get('/calculation/A1234AB/summary/123456')
-      .expect(200)
-      .expect('Content-Type', /html/)
-      .expect(res => {
-        const $ = cheerio.load(res.text)
-        const trancheSelector = $('[data-qa=sds-early-release-tranche]').first()
-        expect(trancheSelector.length).toBe(0)
       })
   })
 })
