+++ server/routes/calculationRoutes.ts
@@ -2,123 +2,24 @@ import { RequestHandler } from 'express'
 import { DateTime } from 'luxon'
 import CalculateReleaseDatesService from '../services/calculateReleaseDatesService'
 import PrisonerService from '../services/prisonerService'
-import logger from '../../logger'
-import { ErrorMessages, ErrorMessageType } from '../types/ErrorMessages'
-import { nunjucksEnv } from '../utils/nunjucksSetup'
 import { FullPageError } from '../types/FullPageError'
 import CalculationSummaryViewModel from '../models/calculation/CalculationSummaryViewModel'
 import UserInputService from '../services/userInputService'
-import { DetailedDate, ManualEntrySelectedDate } from '../@types/calculateReleaseDates/calculateReleaseDatesClientTypes'
+import { DetailedDate } from '../@types/calculateReleaseDates/calculateReleaseDatesClientTypes'
 import CalculationCompleteViewModel from '../models/calculation/CalculationCompleteViewModel'
 import CalculationSummaryPageViewModel from '../models/calculation/CalculationSummaryPageViewModel'
 import { calculationSummaryDatesCardModelFromCalculationSummaryViewModel } from '../views/pages/components/calculation-summary-dates-card/CalculationSummaryDatesCardModel'
-import {
-  ApprovedDateActionConfig,
-  approvedSummaryDatesCardModelFromCalculationSummaryViewModel,
-} from '../views/pages/components/approved-summary-dates-card/ApprovedSummaryDatesCardModel'
-import UserPermissionsService from '../services/userPermissionsService'
+import { approvedSummaryDatesCardModelFromCalculationSummaryViewModel } from '../views/pages/components/approved-summary-dates-card/ApprovedSummaryDatesCardModel'
 import CancelQuestionViewModel from '../models/CancelQuestionViewModel'
 import ConcurrentConsecutiveSentence from '../models/ConcurrentConsecutiveSentencesModel'
-import { ManualJourneySelectedDate } from '../types/ManualJourney'
 
 export default class CalculationRoutes {
   constructor(
     private readonly calculateReleaseDatesService: CalculateReleaseDatesService,
     private readonly prisonerService: PrisonerService,
     private readonly userInputService: UserInputService,
-    private readonly userPermissionsService: UserPermissionsService,
   ) {
     // intentionally left blank
-  }
-
-  public calculationSummary: RequestHandler = async (req, res): Promise<void> => {
-    const { caseloads, token, userRoles } = res.locals.user
-    const { nomsId } = req.params
-    const { callbackUrl } = req.query as Record<string, string>
-    await this.prisonerService.checkPrisonerAccess(nomsId, token, caseloads, userRoles)
-    const calculationRequestId = Number(req.params.calculationRequestId)
-    const detailedCalculationResults = await this.calculateReleaseDatesService.getResultsWithBreakdownAndAdjustments(
-      calculationRequestId,
-      token,
-    )
-    if (detailedCalculationResults.context.prisonerId !== nomsId) {
-      throw FullPageError.notFoundError()
-    }
-    const prisonerDetail = await this.prisonerService.getPrisonerDetail(nomsId, token, caseloads, userRoles)
-    const serverErrors = req.flash('serverErrors')
-    let validationErrors = null
-    if (serverErrors && serverErrors[0]) {
-      validationErrors = JSON.parse(serverErrors[0])
-    }
-    let approvedDates
-    if (
-      req.session.selectedApprovedDates &&
-      req.session.selectedApprovedDates[nomsId] &&
-      req.session.selectedApprovedDates[nomsId].length > 0
-    ) {
-      req.session.selectedApprovedDates[nomsId] = req.session.selectedApprovedDates[nomsId].filter(
-        d => d.completed === true,
-      )
-      approvedDates = this.indexBy(req.session.selectedApprovedDates[nomsId])
-    }
-    if (!req.session.HDCED) {
-      req.session.HDCED = {}
-    }
-    if (!req.session.HDCED_WEEKEND_ADJUSTED) {
-      req.session.HDCED_WEEKEND_ADJUSTED = {}
-    }
-    if (detailedCalculationResults.dates.HDCED) {
-      req.session.HDCED[nomsId] = detailedCalculationResults.dates.HDCED.date
-      const hcedWeekendAdjusted = await this.calculateReleaseDatesService.getNextWorkingDay(
-        detailedCalculationResults.dates.HDCED.date,
-        token,
-      )
-      req.session.HDCED_WEEKEND_ADJUSTED[nomsId] =
-        detailedCalculationResults.dates.HDCED.date != null && hcedWeekendAdjusted != null
-    }
-
-    const model = new CalculationSummaryViewModel(
-      calculationRequestId,
-      nomsId,
-      prisonerDetail,
-      detailedCalculationResults.calculationOriginalData.sentencesAndOffences,
-      false,
-      false,
-      null,
-      detailedCalculationResults.context.calculationReference,
-      false,
-      null,
-      null,
-      null,
-      detailedCalculationResults.calculationBreakdown,
-      detailedCalculationResults.releaseDatesWithAdjustments,
-      validationErrors,
-      false,
-      approvedDates,
-      detailedCalculationResults,
-    )
-    res.render(
-      'pages/calculation/calculationSummary',
-      new CalculationSummaryPageViewModel(
-        model,
-        calculationSummaryDatesCardModelFromCalculationSummaryViewModel(model, false),
-        approvedSummaryDatesCardModelFromCalculationSummaryViewModel(model, true, {
-          nomsId,
-          calculationRequestId,
-        } as ApprovedDateActionConfig),
-        req.session.isAddDatesFlow,
-        callbackUrl || req.originalUrl,
-      ),
-    )
-  }
-
-  private indexBy(dates: ManualJourneySelectedDate[]) {
-    const result = {}
-    dates.forEach(date => {
-      const dateString = `${date.manualEntrySelectedDate.date.year}-${date.manualEntrySelectedDate.date.month}-${date.manualEntrySelectedDate.date.day}`
-      result[date.dateType] = DateTime.fromFormat(dateString, 'yyyy-M-d').toFormat('cccc, dd LLLL yyyy')
-    })
-    return result
   }
 
   private indexApprovedDates(dates: { [key: string]: string } | { [key: string]: DetailedDate }) {
@@ -182,61 +83,6 @@ export default class CalculationRoutes {
         req.originalUrl,
       ),
     )
-  }
-
-  public submitCalculationSummary: RequestHandler = async (req, res): Promise<void> => {
-    const { token, username } = res.locals.user
-    const { nomsId } = req.params
-    const calculationRequestId = Number(req.params.calculationRequestId)
-    const breakdownHtml = await this.getBreakdownFragment(calculationRequestId, token)
-
-    const approvedDates: ManualJourneySelectedDate[] =
-      req.session.selectedApprovedDates != null && req.session.selectedApprovedDates[nomsId] != null
-        ? req.session.selectedApprovedDates[nomsId]
-        : []
-
-    const newApprovedDates: ManualEntrySelectedDate[] = approvedDates.map(d => d.manualEntrySelectedDate)
-    try {
-      const bookingCalculation = await this.calculateReleaseDatesService.confirmCalculation(
-        username,
-        nomsId,
-        calculationRequestId,
-        token,
-        {
-          calculationFragments: {
-            breakdownHtml,
-          },
-          approvedDates: newApprovedDates,
-          isSpecialistSupport: false,
-        },
-      )
-      res.redirect(`/calculation/${nomsId}/complete/${bookingCalculation.calculationRequestId}`)
-    } catch (error) {
-      // TODO Move handling of validation errors from the api into the service layer
-      logger.error(error)
-      if (error.status === 412) {
-        req.flash(
-          'serverErrors',
-          JSON.stringify({
-            messages: [
-              {
-                text: 'The booking data that was used for this calculation has changed, go back to the Check NOMIS Information screen to see the changes',
-                href: `/calculation/${nomsId}/check-information`,
-              },
-            ],
-          } as ErrorMessages),
-        )
-      } else {
-        req.flash(
-          'serverErrors',
-          JSON.stringify({
-            messages: [{ text: 'The calculation could not be saved in NOMIS.' }],
-            messageType: ErrorMessageType.SAVE_DATES,
-          } as ErrorMessages),
-        )
-      }
-      res.redirect(`/calculation/${nomsId}/summary/${calculationRequestId}`)
-    }
   }
 
   public askCancelQuestion: RequestHandler = async (req, res): Promise<void> => {
@@ -339,14 +185,5 @@ export default class CalculationRoutes {
     res.redirect(
       `summary/${preliminaryRequest.calculationRequestId}?callbackUrl=/calculation/${nomsId}/concurrent-consecutive?duration=${sentenceDuration}`,
     )
-  }
-
-  private async getBreakdownFragment(calculationRequestId: number, token: string): Promise<string> {
-    return nunjucksEnv().render('pages/fragments/breakdownFragment.njk', {
-      model: {
-        ...(await this.calculateReleaseDatesService.getBreakdown(calculationRequestId, token)),
-        showBreakdown: () => true,
-      },
-    })
   }
 }
